pyproject.toml
[project]
name = "crm-performance-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "karippery",email = "john.karippery@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "faker (>=37.5.3,<38.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-redis (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
sentry-sdk = "^2.34.1"
django-debug-toolbar = "^6.0.0"
drf-yasg = "^1.21.10"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
pytest-cov = "^6.2.1"

